+++ test/basic/views/Welcome.py
@@ -7,9 +7,12 @@ class Welcome(brueElement):
     def __init__(self):
         super().__init__()
 
+    def test_click(self):
+        print("test!")
+
     def render(self):
         return """
-        <comp-hello name="eseunghwan">
+        <comp-hello name="eseunghwan" :on-click="self.test_click">
             <button>click me!</button>
         </comp-hello>
         """